+++ cupy/_io/formatting.py
