+++ cupy/_io/npz.py
