+++ cupy/_io/text.py
