--- cupy/io/__init__.py
@@ -1,7 +1,0 @@
-# Functions from the following NumPy document
-# https://docs.scipy.org/doc/numpy/reference/routines.io.html
-
-# "NOQA" to suppress flake8 warning
-from cupy.io import formatting  # NOQA
-from cupy.io import npz  # NOQA
-from cupy.io import text  # NOQA
